chore: move function deps function to submodule
